Remove and refactor redundant basin vectorization logic
This commit removes the unused `scripts/make_all_intersecting_polygon.py` and simplifies the logic in the `make_basin.py` module. Also, the process for fixing stream order and interaction with the numpy arrays in `thin_grid.pyx` has been made more efficient with clearer comments explaining the implemented methods. This has improved the clarity of the process' intent and increased the overall performance of the vectorization.